[#30] unitsのfieldを新規ファイルに切り出した
